% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNDmap.R
\name{MNDmap}
\alias{MNDmap}
\title{Map underlying polygon.}
\usage{
MNDmap(map_polygon)
}
\arguments{
\item{map_polygon}{\code{sf} object with the polygon representing the 
territory.}
}
\description{
Map the underlying polygon representing the geographical territory as an
 \code{sf} object.
}
\examples{
txt <- readLines(system.file("extdata/input_files", "map.wkt", package = "simviz"))
map_polygon <- readWKT_as_sf(txt)
MNDmap(map_polygon)

}
