% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNDmap_grid.R
\name{MNDmap_grid}
\alias{MNDmap_grid}
\title{Grid over the map.}
\usage{
MNDmap_grid(MNDmap, grid)
}
\arguments{
\item{MNDmap}{\code{ggplot} object with the polygon representing the 
territory from MNDmap function.}

\item{grid}{\code{stars} object with the grid over the territory.}
}
\description{
Plot the grid over the map (object from MNDmap function)
}
\examples{
txt <- readLines(system.file("extdata/input_files", "map.wkt", package = "simviz"))
map_polygon <- readWKT_as_sf(txt)
MNDmap(map_polygon)

}
