% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNDmap_networkCells.R
\name{MNDmap_networkCells}
\alias{MNDmap_networkCells}
\title{Grid over the map.}
\usage{
MNDmap_networkCells(MNDmap, network_info, coverage)
}
\arguments{
\item{MNDmap}{\code{ggplot} object with the polygon representing the 
territory from MNDmap function with or without the grid.}

\item{grid}{\code{stars} object with the grid over the territory.}
}
\description{
Plot the grid over the map (object from MNDmap function)
}
\examples{
txt <- readLines(system.file("extdata/input_files", "map.wkt", package = "simviz"))
map_polygon <- readWKT_as_sf(txt)
MNDmap(map_polygon)

}
