% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWKT_as_sfc.R
\name{readWKT_as_sfc}
\alias{readWKT_as_sfc}
\title{Read string in WKT format and convert it into an sf geometry list column.}
\usage{
readWKT_as_sfc(
  x,
  crs = NA_integer_,
  precision = 0,
  check_ring_dir = FALSE,
  dim
)
}
\arguments{
\item{crs}{integer or character; coordinate reference system for the geometry
as in function \code{sf::st_as_sfc}}

\item{text}{Character string of WKT}

\item{...}{further arguments passed on internally to function 
\code{sf::st_as_sfc}}

\item{GeoJSON}{logical; if TRUE, try to read geometries from GeoJSON text 
strings geometry as in function \code{sf::st_as_sfc}: see \code{sf::st_crs()}}
}
\description{
Read the input WKT string, parse it, and convert it into an sf geometry list 
column.
}
\examples{
txt <- readLines(system.file("extdata/input_files", "map.wkt", package = "simviz"))
map_polygon_sfc <- readWKT_as_sfc(txt)
class(map_polygon_sfc)

}
