% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplot.R
\name{simplot}
\alias{simplot}
\title{Visualization of simulator outputs.}
\usage{
simplot(
  simElem.list,
  control.map = list(),
  control.grid = list(),
  control.network = list(),
  control.coverage = list(),
  control.individuals = list(animate = FALSE)
)
}
\arguments{
\item{control.map}{a list with the visualization parameters for the map object:
size, ...}

\item{control.grid}{a list with the visualization parameters for the grid object:
fill, ...}

\item{control.network}{a list with the visualization parameters for the network object.}

\item{control.coverage}{a list with the visualization parameters for the coverage object.}

\item{control.individuals}{a list with the visualization parameters for the individuals object.}

\item{simElements}{a list with the output objects from the simulator created 
by the \code{create_simElements} function.}
}
\description{
Make the visualization of the objects obtained from the MNO data simulator.
}
\examples{
filename_map      <- system.file("extdata/input_files", "map.wkt", package = "simviz")
filename_network  <- c(csv = system.file("extdata/output_files", "antennas.csv", package = "simviz"),
                      xml = system.file("extdata/metadata/output_files", "antennas_dict.xml", package = "simviz"))

filename_coverage <- c(csv = system.file("extdata/output_files", "AntennaCells_MNO1.csv", package = "simviz"),
                      xml = system.file("extdata/metadata/output_files", "AntennaCells_dict.xml", package = "simviz"))

filename_grid     <- c(csv = system.file("extdata/output_files", "grid.csv", package = "simviz"),
                      xml = system.file("extdata/metadata/output_files", "grid_dict.xml", package = "simviz"))

filename_individ  <- c(csv = system.file("extdata/output_files", "persons.csv", package = "simviz"),
                      xml = system.file("extdata/metadata/output_files", "persons_dict.xml", package = "simviz"))   

filenames <- list(
 map                = filename_map,
 network_parameters = filename_network,
 coverage_cells     = filename_coverage,
 grid               = filename_grid,
 individuals        = filename_individ)

simElem.list <- create_simElements(filenames, crs = 2062)

simplot(simElem.list, control.map = list(size = 1.5),
control.grid = list(fill = NA),
control.coverage = list(alpha = 0.1, fill = "power"),
control.network = list(label = "Antenna ID", nudge_x = 1, nudge_y = 1),
control.individuals = list(t = 0, animate = FALSE))

}
