% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplot_territory.R
\name{simplot_territory}
\alias{simplot_territory}
\title{Visualization of simulator inputs and outputs: Territory information.}
\usage{
simplot_territory(simData.list, aggre_level, aggre_name, plot_title)
}
\arguments{
\item{simData.list}{a list with the objects from the simulator created 
by the \code{read_simData} function from simutils R package.}

\item{aggre_level}{numeric, until now 1 for regions and 2 for subregions.}

\item{aggre_name}{string with the name of the aggregation level.}

\item{plot_title}{string with the general title of the plot.}
}
\value{
It returns an object of class \code{ggplot} with the graph.
}
\description{
Make the visualization of the objects obtained from the MNO data simulator 
after being managed by simutils package.
}
\examples{
filename_map      <- c(xml= system.file("extdata/input_files", "map.xml", package = "simutils"),  xsd= '')

filename_network  <- c(csv= system.file("extdata/output_files/antennas.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/antennas_dict.xml", package = "simutils"))

filename_signal <- c(csv= system.file("extdata/output_files/SignalMeasure_MNO1.csv", package = "simutils"),  
 xml= system.file("extdata/metadata/output_files/SignalMeasure_dict.xml", package = "simutils"))

filename_coverage <- c(csv= system.file("extdata/output_files", "AntennaCells_MNO1.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/AntennaCells_dict.xml", package = "simutils"))

filename_grid <- c(csv= system.file("extdata/output_files/grid.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/grid_dict.xml", package = "simutils")) 

filename_individ <- c(csv= system.file("extdata/output_files/persons_dash.csv", package = "simutils"),
 xml= system.file("extdata/metadata/output_files/persons_dash_dict.xml", package = "simutils"))   

filenames <- list(map                = filename_map,  
                 network_parameters = filename_network,
                 signal             = filename_signal,
                 coverage_cells     = filename_coverage,
                 grid               = filename_grid,
                 individuals        = filename_individ)

simData <- read_simData(filenames, crs = 2062)

simplot_territory(simData, 
                 aggre_level = 1, 
                 aggre_name = "Regions", 
                 plot_title = 'Territorial map for simulation')


}
