% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_attrs2dt.R
\name{xml_attrs2dt}
\alias{xml_attrs2dt}
\alias{xmlTypes2RTypes}
\title{Create a data.table from a file metadata xml.}
\usage{
xmlTypes2RTypes(xmlType)
}
\arguments{
\item{xmlNode}{xml node of the input xml file}

\item{dataset}{character vector of length 1 with the dataset name}
}
\description{
Create a data.table from a file metadata xml.
}
\examples{
xmlname <- system.file("extdata/metadata/output_files", "antennas_dict.xml", package = "simviz") 
xml_attrs2dt(xmlname, 'antennas')

xmlname <- system.file("extdata/metadata/input_files", "persons_dict.xml", package = "simviz") 
xml_attrs2dt(xsdname, 'individuals')

}
