% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsd_attrs2dt.R
\name{xsd_attrs2dt}
\alias{xsd_attrs2dt}
\alias{xsd_node_attrs2dt}
\title{Create a data.table from a file metadata xsd.}
\usage{
xsd_node_attrs2dt(xsdNode, dataset)
}
\arguments{
\item{xsdNode}{xml node of the input xsd file}

\item{dataset}{character vector of length 1 with the dataset name}
}
\description{
Create a data.table from a file metadata xsd.
}
\examples{
xsdname <- system.file("extdata", "antennas_dict.xsd", package = "simviz") 
xsd_attrs2dt(xsdname, 'antennas')
}
