% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWKT_as_sf.R
\name{readWKT_as_sf}
\alias{readWKT_as_sf}
\title{Read string in WKT format and convert it into an sf object.}
\usage{
readWKT_as_sf(text, id = NULL, p4s = NULL, ...)
}
\arguments{
\item{text}{Character string of WKT}

\item{id}{Character vector of unique ids to label geometries as in function
\code{rgeos::readWKT}. Length must match the number of subgeometries in the
WKT}

\item{p4s}{Either a character string or an object of class \code{CRS} as in
the function \code{rgeos::readWKT}}

\item{...}{passed on to \code{sf::st_as_sf}, might included named arguments
crs or precision}
}
\description{
Read the input WKT string, parse it, and convert it into an sf object.
}
\examples{
txt <- readLines(system.file("extdata", "map.wkt", package = "simviz"))
map_polygon <- readWKT_as_sf(txt)
class(map_polygon)

}
